fetch/metadata/generated/global-properties-sec-fetch-site.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <meta name="timeout" content="long">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTP</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, navigate) {
    const win = window.open();
    const fullPath = url;

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });

        navigate(win, fullPath);
      })
      .then(() => win.close());
  }

  const responseParams = {
    mime: 'text/html',
    body: `<script>opener.postMessage('done', '*')</${''}script>`
  };

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location = path;
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'HTTPS downgrade (header not sent) - location');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.href = path;
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'HTTPS downgrade (header not sent) - location.href');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.assign(path);
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'HTTPS downgrade (header not sent) - location.assign');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.replace(path);
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'HTTPS downgrade (header not sent) - location.replace');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location = path;
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS upgrade - location');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.href = path;
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS upgrade - location.href');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.assign(path);
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS upgrade - location.assign');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.replace(path);
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS upgrade - location.replace');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location = path;
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS downgrade-upgrade - location');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.href = path;
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS downgrade-upgrade - location.href');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.assign(path);
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS downgrade-upgrade - location.assign');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin'], responseParams);

    const navigate = (win, path) => {
      win.location.replace(path);
    };
    return induceRequest(url, navigate)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS downgrade-upgrade - location.replace');
  </script>
  </body>
</html>
